chore: extract CyberpunkCity component to its own file
Extracting the long hardcoded CyberpunkCity component from ImprovedGame.tsx
into src/components/CyberpunkCity.tsx to improve maintainability and
readability.
